+++ Controller/ProfilesController.php
@@ -120,7 +120,7 @@ class ProfilesController extends MeCmsAppController {
 	 * @uses RecaptchaComponent::check()
 	 * @uses TokenComponent::create()
 	 */
-	public function forgot_password() {		
+	public function forgot_password() {
 		//Redirects if the user is already logged in
 		$this->redirectIfLogged();
 		
@@ -133,7 +133,7 @@ class ProfilesController extends MeCmsAppController {
 		if($this->request->is('post') || $this->request->is('put')) {
 			//Sets data
 			$this->User->set($this->request->data);
-			//Email should not be unique, no change required
+			//Email should not be unique, removes validation rule
 			unset($this->User->validate['email']['isUnique']);
 			
 			//Checks for reCAPTCHA, if requested
@@ -171,6 +171,68 @@ class ProfilesController extends MeCmsAppController {
 		}
 		
 		$this->set('title_for_layout', __d('me_cms', 'Forgot your password?'));
+		$this->layout = 'MeCms.users';
+	}
+	
+	/**
+	 * Resends the activation mail.
+	 * @uses config
+	 * @uses RecaptchaComponent::error
+	 * @uses redirectIfLogged()
+	 * @uses RecaptchaComponent::check()
+	 * @uses TokenComponent::create()
+	 */
+	public function resend_activation() {
+		//Redirects if the user is already logged in
+		$this->redirectIfLogged();
+		
+		//Checks if signup is enabled and if accounts will be enabled by the user via email
+		if(!$this->config['users']['signup'] && $this->config['users']['activation'] === 1) {
+			$this->Session->flash(__d('me_cms', 'Disabled'), 'error');
+			$this->redirect($this->Auth->loginAction);
+		}
+		
+		if($this->request->is('post') || $this->request->is('put')) {
+			//Sets data
+			$this->User->set($this->request->data);
+			//Email should not be unique, removes validation rule
+			unset($this->User->validate['email']['isUnique']);
+			
+			//Checks for reCAPTCHA, if requested
+			if($this->config['security']['recaptcha'] && !$this->Recaptcha->check()) {
+				$this->Session->flash($this->Recaptcha->error, 'error');
+			}
+			elseif($this->User->validates()) {
+				//Gets the user
+				$user = $this->User->find('pending', array(
+					'conditions'	=> array('email' => $email = $this->request->data['User']['email']),
+					'fields'		=> array('id', 'email', 'full_name'),
+					'limit'			=> 1
+				));
+			
+				if(!empty($user)) {
+					//Gets the token and the url to activate account
+					$token = $this->Token->create($email = $user['User']['email'], array('type' => 'signup', 'user_id' => $id = $user['User']['id']));
+					$url = Router::url(am(array('controller' => 'profiles', 'action' => 'activate_account'), compact('id', 'token')), TRUE);
+
+					//Sends email
+					$this->Email->to(array($email => $full_name = $user['User']['full_name']));
+					$this->Email->subject(__d('me_cms', 'Activate your account'));
+					$this->Email->template('signup');
+					$this->Email->set(compact('full_name', 'url'));
+					$this->Email->send();
+
+					$this->Session->flash(__d('me_cms', 'We send you an email to activate your account'));
+					$this->redirect($this->Auth->loginAction);
+				}
+				else
+					$this->Session->flash(__d('me_cms', 'No account found'), 'error');
+			}
+			else
+				$this->Session->flash(__d('me_cms', 'The form has not been filled in correctly, try again'), 'error');
+		}
+		
+		$this->set('title_for_layout', __d('me_cms', 'Resend activation email'));
 		$this->layout = 'MeCms.users';
 	}